requirements.txt
# PaperFlow v2.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
resend==2.23.0
pypdf==5.1.0
PyMuPDF==1.27.2
pymupdf4llm==1.27.2.1
